Used user_agent to render mobile & PC seperately
The layout of this site was designed in a way that creating
the mobile and desktop views using pure css/js would have been
very complex due to the differing layouts. By using django-user_agent
I can discern whether the user is on mobile or desktop and render
templates unique to that viewport.