Changed file structure to include "includes" for html files
The previous structure had two seperate base html files that python
then loaded depending on viewport size. This worked but is not
a production level way of writing code.

Instead by adding an includes folder and django logic to the base
file you can clearly see what happens behind the curtain when
looking at the code. It also makes it easier to re-write, edit or add
to the seperate files without breaking the main base template. As well
as reducing duplicated code.