Customised admin page on website
Due to the being a live website, a website that is going to be
maintained by someone other than me mostly there needed to be a better
way to view the admin page where the owner will add products. So by
adding extra fields for viewing and sorting products by sku it will
make maitenance easier and faster.