Fixed bug where only first price was changeable
When I created the vat toggler I used id's instead of classes
this was fine when only one price was being generated. But when
a list of products was shown only the first correctly changed prices
on click.

To solve I changed the incVat and excVat id's to classes in the
js function and html.